fix(desktop-app): upload モジュールの循環依存を解消
Image 型を types.ts に抽出し、api.ts → local-api.ts → api.ts の
循環参照を解消。
